Dodanie ruchomego punktu środkowego

+++ src/seventh/tools/PickTool.py
@@ -1,4 +1,5 @@
 import tkinter as tk
+from statistics import mean
 from typing import Dict, Callable
 
 from src.seventh.tools.AbstractTool import AbstractTool
@@ -8,6 +9,7 @@ class PickTool(AbstractTool):
 
     def __init__(self, *args, main_window: 'MainWindow' = None, **kwargs) -> None:
         self._moved_object: int | None = None
+        self.center_point: int | None = None
         self._first_click_cords: tuple[int] | list[int] | None = None
         super().__init__(*args, main_window=main_window, **kwargs)
 
@@ -28,9 +30,10 @@ class PickTool(AbstractTool):
         cords = event.x, event.y
         ids = main_window.canvas.find_overlapping(*cords, *cords)
 
-
         intersection_with_box_points = main_window.items_to_be_deleted_at_changing_tools.intersection(ids)
         if intersection_with_box_points:
+            self._first_click_cords = cords
+            self._moved_object = self.center_point
             return d
 
         identifier = None
@@ -51,15 +54,18 @@ class PickTool(AbstractTool):
     def reset(self, main_window):
         self.uncheck_item(main_window)
         main_window.delete_items_to_be_deleted()
+        self.center_point = None
 
     def generate_box_points(self, identifier, main_window):
         raw_coords = main_window.canvas.coords(identifier)
         xs, ys = raw_coords[::2], raw_coords[1::2]
 
         min_x = min(xs)
+        avg_x = mean(xs)
         max_x = max(xs)
 
         min_y = min(ys)
+        avg_y = mean(ys)
         max_y = max(ys)
 
         for x in [min_x, max_x]:
@@ -68,6 +74,11 @@ class PickTool(AbstractTool):
                     .create_oval(x - 5, y - 5,
                                  x + 5, y + 5,
                                  fill="blue")
+        self.center_point = main_window.canvas.create_oval(avg_x - 5, avg_y - 5,
+                                                           avg_x + 5, avg_y + 5,
+                                                           fill="yellow")
+
+        main_window.items_to_be_deleted_at_changing_tools = self.center_point
 
     def check_item(self, ids, main_window):
         main_window.checked_item = ids
@@ -85,8 +96,9 @@ class PickTool(AbstractTool):
             self._first_click_cords = cords
             main_window.canvas.move(self._moved_object, *delta)
 
-            main_window.delete_items_to_be_deleted()
-            self.generate_box_points(self._moved_object, main_window)
+            if self._moved_object not in main_window.items_to_be_deleted_at_changing_tools:
+                main_window.delete_items_to_be_deleted()
+                self.generate_box_points(self._moved_object, main_window)
         return super().after_first_click_motion(*args, **kwargs)
 
     def double_click_in_canvas(self, main_window: 'MainWindow', event: tk.Event, *args, **kwargs):
